# Chaining hash table for repository.Index