--- a/changelog/unreleased/pull-2781
+++ b/changelog/unreleased/pull-2781
@@ -1,6 +1,8 @@
 Enhancement: Reduce memory consumption of in-memory index
 
 We've improved how the index is stored in memory.
-This change reduces memory usage for large repositories by about 30-40%.
+This change can reduce memory usage for large repositories by up to 50%
+(depending on the operation).
 
 https://github.com/restic/restic/pull/2781
+https://github.com/restic/restic/pull/2812