What is the purpose of this change? What does it change?
--------------------------------------------------------

This implements a custom chaining hash table for repository.Index.

Index merging (#2799) should be easy to add to this PR.

Benchmark results:

```
name                                     old time/op    new time/op    delta
PackerManager-8                             180ms ± 1%     186ms ± 0%   +3.15%  (p=0.004 n=6+5)
DecodeIndex-8                              13.7µs ± 2%    16.6µs ± 3%  +21.43%  (p=0.002 n=6+6)
IndexHasUnknown-8                          43.3ns ± 5%    56.4ns ±20%  +30.23%  (p=0.004 n=6+6)
IndexHasKnown-8                            49.4ns ± 3%    60.0ns ±21%  +21.52%  (p=0.002 n=6+6)
IndexAlloc-8                                812ms ± 2%     952ms ± 1%  +17.21%  (p=0.002 n=6+6)
MasterIndexLookupSingleIndex-8              122ns ± 1%     133ns ± 8%   +8.58%  (p=0.002 n=6+6)
MasterIndexLookupMultipleIndex-8            380ns ± 2%     450ns ± 3%  +18.26%  (p=0.004 n=6+5)
MasterIndexLookupSingleIndexUnknown-8      69.6ns ± 7%    71.0ns ±11%     ~     (p=0.554 n=6+6)
MasterIndexLookupMultipleIndexUnknown-8     318ns ± 1%     348ns ±10%   +9.53%  (p=0.002 n=6+6)
SaveAndEncrypt-8                            101ns ± 2%     101ns ± 1%     ~     (p=1.000 n=6+6)
LoadTree-8                                  337µs ± 5%     345µs ± 1%     ~     (p=0.394 n=6+6)
LoadBlob-8                                 5.80ms ± 2%    5.67ms ± 0%   -2.16%  (p=0.009 n=6+5)
LoadAndDecrypt-8                           7.15ms ± 4%    7.09ms ± 2%     ~     (p=0.937 n=6+6)
LoadIndex-8                                33.0ms ± 2%    33.7ms ± 2%   +2.12%  (p=0.015 n=6+6)

name                                     old speed      new speed      delta
PackerManager-8                           292MB/s ± 1%   283MB/s ± 0%   -3.06%  (p=0.004 n=6+5)
SaveAndEncrypt-8                         41.4TB/s ± 2%  41.6TB/s ± 1%     ~     (p=0.818 n=6+6)
LoadBlob-8                                173MB/s ± 2%   176MB/s ± 0%   +2.19%  (p=0.009 n=6+5)
LoadAndDecrypt-8                          140MB/s ± 4%   141MB/s ± 2%     ~     (p=0.937 n=6+6)

name                                     old alloc/op   new alloc/op   delta
PackerManager-8                            91.5kB ± 0%    91.5kB ± 0%     ~     (p=0.818 n=6+5)
IndexAlloc-8                                786MB ± 0%     464MB ± 0%  -40.92%  (p=0.002 n=6+6)
SaveAndEncrypt-8                            1.00B ± 0%     1.33B ±50%     ~     (p=0.455 n=6+6)

name                                     old allocs/op  new allocs/op  delta
PackerManager-8                             1.41k ± 0%     1.41k ± 0%     ~     (p=0.119 n=6+6)
IndexAlloc-8                                 977k ± 0%     1008k ± 0%   +3.22%  (p=0.002 n=6+6)
SaveAndEncrypt-8                             0.00           0.00          ~     (all equal)
```

Was the change discussed in an issue or in the forum before?
------------------------------------------------------------

Alternative to #2811, with better performance on the benchmarks.

Checklist
---------

- [x] I have read the [Contribution Guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches)
- [x] I have enabled [maintainer edits for this PR](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork)
- [ ] I have added tests for all changes in this PR
- [ ] I have added documentation for the changes (in the manual)
- [ ] There's a new file in `changelog/unreleased/` that describes the changes for our users (template [here](https://github.com/restic/restic/blob/master/changelog/TEMPLATE))
- [x] I have run `gofmt` on the code in all commits
- [ ] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits)
- [ ] I'm done, this Pull Request is ready for review