@greatroar 2020-06-28
Here's a quick real-world test using restic cat, as suggested by @aawsome. $id is the id of an existing tree blob in a 700GiB repo (local disk).

Current master:
```
$ /usr/bin/time ./restic cat blob $id >/dev/null 
12.77user 0.32system 0:04.39elapsed 298%CPU (0avgtext+0avgdata 705296maxresident)k
0inputs+0outputs (0major+125964minor)pagefaults 0swaps
$ /usr/bin/time ./restic -cat blob $id >/dev/null 
12.70user 0.35system 0:04.01elapsed 325%CPU (0avgtext+0avgdata 717884maxresident)k
0inputs+0outputs (0major+127844minor)pagefaults 0swaps
$ /usr/bin/time ./restic cat blob $id >/dev/null 
12.90user 0.34system 0:04.10elapsed 322%CPU (0avgtext+0avgdata 727872maxresident)k
0inputs+0outputs (0major+139900minor)pagefaults 0swaps
```

This PR:

```
$ /usr/bin/time ./restic cat blob $id >/dev/null 
13.77user 0.35system 0:04.38elapsed 322%CPU (0avgtext+0avgdata 569968maxresident)k
0inputs+0outputs (0major+136284minor)pagefaults 0swaps
$ /usr/bin/time ./restic cat blob $id >/dev/null 
13.60user 0.28system 0:03.99elapsed 348%CPU (0avgtext+0avgdata 533716maxresident)k
0inputs+0outputs (0major+123454minor)pagefaults 0swaps
$ /usr/bin/time ./restic cat blob $id >/dev/null 
13.91user 0.31system 0:04.07elapsed 349%CPU (0avgtext+0avgdata 570852maxresident)k
0inputs+0outputs (0major+137047minor)pagefaults 0swaps
```

#2811 at 64ecc669b6720fe1160731c6be895a1810347938:

```
$ /usr/bin/time ./restic -p ~/.restic/pw cat blob $id >/dev/null 
13.34user 0.22system 0:04.87elapsed 278%CPU (0avgtext+0avgdata 467384maxresident)k
0inputs+0outputs (0major+105581minor)pagefaults 0swaps
$ /usr/bin/time ./restic -p ~/.restic/pw cat blob $id >/dev/null 
13.22user 0.27system 0:04.18elapsed 322%CPU (0avgtext+0avgdata 478512maxresident)k
0inputs+0outputs (0major+109278minor)pagefaults 0swaps
$ /usr/bin/time ./restic -p ~/.restic/pw cat blob $id >/dev/null 
13.27user 0.24system 0:04.15elapsed 325%CPU (0avgtext+0avgdata 468200maxresident)k
0inputs+0outputs (0major+104110minor)pagefaults 0swaps
```

Here, this PR so far achieves a 23% reduction in average memory use, while #2811 gets a 35% reduction. There's a few constants left to tune, I'll see if can lower.