@greatroar 2020-06-28
Nope, tuning constants doesn't make much of a difference. I tried again on a smaller, 210GiB repo, and there it's 32% for #2811, 18% for this PR.