@greatroar 2020-06-28
Following @aawsome's advice, I reorganised the duplicates and increased the maximum load factor to make more efficient use of memory. In the first restic cat tests, I now get a 32% reduction and I support I can get up to the level of #2811 by increasing it further.