@greatroar 2020-06-28
Test coverage for indexmap.go is 98.1%, or one return statement not covered.