@aawsome 2020-06-28
I would also add a changelog, as this is another massive improvement with respect to memory usage!