@greatroar 2020-06-28
I intend to merge the changelog entry with the one for #2781, because to a user, they're the same thing. Is that ok with you?