@aawsome 2020-06-28
> I intend to merge the changelog entry with the one for #2781, because to a user, they're the same thing. Is that ok with you?

Sure - I think it is good idea to tell the overall improvement to users.