@greatroar 2020-06-28
Re: test coverage,

```
go test -coverprofile=cov
go tool cover -html cov
```