@aawsome 2020-07-03
> ```go
> type bucket struct {
> 	hash  uint
> 	entry *indexEntry // no longer needs a next pointer
> }
> ```
> 
> Say we set the maximum load factor at 80% and grow in powers of two. Then the table will be 60% full on average and the worst case overhead will be 16*(1+2/3)*n = 26.7n bytes. Dropping the next pointer saves eight bytes, and we pay 74.7 bytes per entry.

No need to save the hash as it can be easily computed.
If you save `*indexEntry` in your bucket and use probing, you gain 8 bytes per index entry. In your actual implementation your buckets need between 2 and 4 bytes per index entry. So if you could use a probing table with load factor between 2/3 and 4/5 (that should be the maximum load factor!), you will get the identical memory usage. However, if you grow a hash table with load factor 80% by doubling its buckets, it will result in a hash table with load factor 40%. Then the buckets will use 5/2 * 8 bytes = 20 bytes per entry.

And that is the main drawback of hashmaps with load factor smaller than 1 - you get memory spikes when growing...