@greatroar 2020-07-03
> No need to save the hash as it can be easily computed.

The hash is saved so four of them will be in the L1 cache before we follow a single pointer. That's also the only benefit of this proposed scheme.