@greatroar 2020-07-17
I just reran the latest benchmarks (fast createRandomIndex, proper size for DecodeIndex*) with entryAllocBatch = 8 as old and 256 as new. These are the results:

```
name                                            old time/op    new time/op    delta
IndexAlloc-8                                       688ms ± 8%     656ms ± 5%   -4.68%  (p=0.006 n=20+20)
IndexAllocParallel-8                               306ms ±16%     233ms ±14%  -23.73%  (p=0.000 n=20+20)
IndexMapHash-8                                    3.61µs ± 0%    3.61µs ± 1%     ~     (p=0.484 n=9+10)
PackerManager-8                                    181ms ± 1%     179ms ± 0%   -0.77%  (p=0.000 n=10+9)
DecodeIndex-8                                      4.34s ± 0%     4.34s ± 1%     ~     (p=0.604 n=9+10)
DecodeIndexParallel-8                              4.34s ± 0%     4.34s ± 1%     ~     (p=0.842 n=10+9)
IndexHasUnknown-8                                 49.4ns ±12%    53.0ns ±16%     ~     (p=0.196 n=10+10)
IndexHasKnown-8                                   54.3ns ±17%    51.7ns ± 9%     ~     (p=0.210 n=10+10)
MasterIndexLookupSingleIndex-8                     144ns ± 5%     138ns ± 3%   -4.38%  (p=0.007 n=10+9)
MasterIndexLookupMultipleIndex-8                   419ns ± 6%     424ns ± 4%     ~     (p=0.424 n=10+10)
MasterIndexLookupSingleIndexUnknown-8             73.5ns ± 9%    70.7ns ± 8%     ~     (p=0.305 n=10+9)
MasterIndexLookupMultipleIndexUnknown-8            351ns ±10%     353ns ± 3%     ~     (p=0.328 n=9+9)
MasterIndexLookupParallel/known,indices=5-8        859ns ± 2%     879ns ± 5%   +2.30%  (p=0.011 n=10+10)
MasterIndexLookupParallel/unknown,indices=5-8      278ns ± 2%     295ns ± 1%   +6.31%  (p=0.000 n=10+8)
MasterIndexLookupParallel/known,indices=10-8      1.17µs ± 1%    1.14µs ± 1%   -2.57%  (p=0.000 n=10+10)
MasterIndexLookupParallel/unknown,indices=10-8     261ns ± 1%     253ns ± 1%   -3.07%  (p=0.000 n=10+10)
MasterIndexLookupParallel/known,indices=20-8      1.29µs ± 0%    1.21µs ± 2%   -5.71%  (p=0.000 n=7+10)
MasterIndexLookupParallel/unknown,indices=20-8     564ns ± 1%     533ns ± 1%   -5.63%  (p=0.000 n=10+10)
SaveAndEncrypt-8                                  98.8ns ± 1%    98.5ns ± 0%     ~     (p=0.348 n=9+9)
LoadTree-8                                         577µs ± 1%     575µs ± 0%   -0.50%  (p=0.011 n=9+9)
LoadBlob-8                                        5.70ms ± 2%    5.72ms ± 2%     ~     (p=0.280 n=10+10)
LoadAndDecrypt-8                                  5.93ms ± 0%    5.93ms ± 0%     ~     (p=0.243 n=9+10)
LoadIndex-8                                       24.7ms ± 0%    24.6ms ± 1%   -0.50%  (p=0.003 n=10+10)

name                                            old alloc/op   new alloc/op   delta
IndexAlloc-8                                       401MB ± 0%     401MB ± 0%   +0.00%  (p=0.000 n=18+19)
IndexAllocParallel-8                               400MB ± 0%     400MB ± 0%   +0.00%  (p=0.000 n=20+20)
IndexMapHash-8                                     0.00B          0.00B          ~     (all equal)
PackerManager-8                                   91.6kB ± 0%    91.6kB ± 0%     ~     (p=0.291 n=9+9)
SaveAndEncrypt-8                                   1.00B ± 0%     1.00B ± 0%     ~     (all equal)

name                                            old allocs/op  new allocs/op  delta
IndexAlloc-8                                       1.12M ± 0%     0.91M ± 0%  -18.79%  (p=0.000 n=19+18)
IndexAllocParallel-8                               1.12M ± 0%     0.91M ± 0%  -18.79%  (p=0.000 n=20+20)
IndexMapHash-8                                      0.00           0.00          ~     (all equal)
PackerManager-8                                    1.41k ± 0%     1.41k ± 0%     ~     (p=0.173 n=10+10)
SaveAndEncrypt-8                                    0.00           0.00          ~     (all equal)

name                                            old speed      new speed      delta
IndexMapHash-8                                  1.13GB/s ± 0%  1.13GB/s ± 1%     ~     (p=0.497 n=9+10)
PackerManager-8                                  291MB/s ± 1%   294MB/s ± 0%   +0.77%  (p=0.000 n=10+9)
SaveAndEncrypt-8                                42.5TB/s ± 1%  42.6TB/s ± 0%     ~     (p=0.297 n=9+9)
LoadBlob-8                                       176MB/s ± 2%   175MB/s ± 2%     ~     (p=0.280 n=10+10)
LoadAndDecrypt-8                                 169MB/s ± 0%   169MB/s ± 0%     ~     (p=0.242 n=9+10)
```

In summary, constructing indexes gets faster, while lookups get a bit slower. Still nothing that merged indexes couldn't solve, I think, given that the loss is mainly in the lookups of unknown blobs. The overallocation is statistically significant but still tiny.