@MichaelEischer 2020-07-18
I've also run a few of the lookup benchmarks with `-benchtime 10s` and for those there's nearly no difference between a batchsize of 8 or 256, with slightly better performance for a batchsize of 256.

```
name                                            old time/op  new time/op  delta
MasterIndexLookupSingleIndex-8                   134ns ± 6%   131ns ± 6%    ~     (p=0.173 n=20+20)
MasterIndexLookupMultipleIndex-8                 376ns ± 7%   364ns ± 7%  -3.26%  (p=0.011 n=19+19)
MasterIndexLookupSingleIndexUnknown-8           60.2ns ±11%  60.5ns ±11%    ~     (p=0.700 n=20+20)
MasterIndexLookupMultipleIndexUnknown-8          285ns ± 7%   287ns ±13%    ~     (p=0.808 n=19+20)
MasterIndexLookupParallel/known,indices=5-8      342ns ± 5%   340ns ± 5%    ~     (p=0.440 n=20+20)
MasterIndexLookupParallel/unknown,indices=5-8    232ns ± 8%   232ns ± 6%    ~     (p=0.995 n=20+20)
MasterIndexLookupParallel/known,indices=10-8     684ns ± 7%   673ns ± 4%    ~     (p=0.122 n=20+20)
MasterIndexLookupParallel/unknown,indices=10-8   318ns ± 4%   316ns ± 9%    ~     (p=0.399 n=20+19)
MasterIndexLookupParallel/known,indices=20-8     720ns ± 8%   707ns ± 9%    ~     (p=0.168 n=20+19)
MasterIndexLookupParallel/unknown,indices=20-8   487ns ± 4%   490ns ± 4%    ~     (p=0.285 n=20+20)
```

So I'd say we use `256` as batch size, which also has the benefit of a slightly lower memory usage.
@greatroar Do you want to rebase / squash / clean up some of the commits before merging? Or did I miss any points which still need discussion?